[指摘箇所修正]addClipアクションにロジック移動
